backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
mediapipe==0.10.9
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
python-socketio==5.11.0
aiofiles==23.2.1
groq==0.4.2
python-dotenv==1.0.1
